requirements.txt
langchain
langchain-community
langchain-core
langchain-openai
openai
weaviate-client
python-dotenv
pypdf
faiss-cpu